Add expanded customization section with examples
Adds a separate section for theming with examples I've taken from cg-style. Also includes a section on where to find customization variables, although they are all in the `core/variables` file. It might make more sense to make it less about finding and more about what are the important things to change.