Add links from metric discussion to metrics
To ensure it's easy for the user to navigation around to different
metrics as they are being discussed

TODO
Theres currently a visual problem where the anchor links are offset
by 20 or so pixels. So when you arrive at an anchor link after clickin
the link, you arrive too far down and don't see the section title.
Strangely, this only happens when clicking links from the top of the
page, not when you load the page with the anchor link URL.