Fix code of conduct URL in CONTRIBUTING.md.
Unfortunately, the existing CoC link 404's for public contributors due to
the fact that the 18F/code-of-conduct repository is currently private.

For more details on this issue, see:

  https://github.com/18F/code-of-conduct/issues/4

This PR changes the link to point at 18F's publicly-visible copy of
the latest OGC-approved Code of Conduct, located at:

  https://18f.gsa.gov/code-of-conduct/

This page also contains a link to the CoC repository, so that
contributors can still comment on or suggest changes to it
there (well, at least it's once made public again).

It's assumed that once the new CoC becomes approved by OGC, the
aforementioned copy will be updated as needed, so the
link will still be accurate.