3.0 changes needed from Devs_How to use UsWDS
I've revised the summary text  under "Developers," "Designers," and "USWDS project wiki."

As we prepare for the release of 3.0, the following issue is outstanding:

From this page, we link to our project wiki, in which v1 is referenced under current initiatives.
We should remove this link (at least temporarily, though maybe permanently since we aren't updating this content and it's weird it's presented the same as our user guidance). If we keep the link, we should update the wiki.