Updating link for Colors section
On the Data Visualization page, update the link in the Colors section that points to the Nightingale article on Medium. The link should now point to: https://nightingaledvs.com/how-to-choose-colors-for-your-data-visualizations/

Link provided by Amanda Makulec from Nightingale, who pointed out that the previous link to Medium sometimes became paywalled for users. Nightingale has migrated to a standalone platform that does not have a paywall.