Proof ds.d.gov links with relative links
In htmlproofer, replace designsystem.digital.gov links with a relative link so canonical links in the head resolve properly, even with pages that aren't yet published

Ignore designsystem.digital.gov urls in proofer

Why? Links to internal docs are typically set with local urls like `/components/accordion/`. If we see a url that points to ds.d.g, it probably means it's automatically generated, like — specifically — a canonical link in in the head. We don't want proofer to fail every time we add a new page, since in development, the canonical link does not yet exist