Update home.md
Discussed with content and a11y team.

ANDI is flagging that the alt text in each of these images is redundant. Using "image remover" browser extention to Chrome one can see the page without the images and tell that those screen shots aren't critical for figuring out the purpose of those links

We propose removing alt text, as the images are decorative. The links to the sites are what's important.

We also propose changing the title on that section from "Showcase" to "Sites built with USWDS". It's clearer than Showcase.